SOL-561 moved tests to subfolder.
