+++ cypress/fixtures/suite/api/backend/abstract-products/dynamic-abstract-products.json
