+++ cypress/fixtures/suite/api/backend/abstract-products/static-abstract-products.json
