+++ cypress/fixtures/suite/api/backend/products/dynamic-products.json
