+++ cypress/fixtures/suite/api/backend/products/static-products.json
